+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- ** Release 0.1.3 December 3rd 2017 **
+              - Code modularity improvements: added methods get_optimization_quality, get_optimization_setting & get_adjusted_size_params to class HDEV_OPTIMG_Optimize.
+              - Fix: jpeg quality was too low when image was cropped, scaled or rotated from the edit image screen.
+              
 - ** Release 0.1.1 December 1st 2017 **
               - Minor fix: wp_editor_set_quality filter set to 100 to disable WP compression added only if optimization is active.